refactor(agent): split run orchestration domains
